Fix the Failed to open adapter windows problem
Opening in non-shared mode fails on many systems due to
presence of RGB control software (RGB Fusion, etc) which
apparently open the adapter in shared mode.

Once a process has a shared mode handle, opening in exclusive mode
is not possible and causes an error.

I think it is still a good thing to open the device in exclusive
mode if it works. But if it fails, fall-back to a shared open.